Split out bootstrap.sh into scripts directory
Environment variables are difficult to work with and not really
neccessary for most of the template. This splits those functions into
separate scripts that can be called indivually in the provisioner
blocks.

This also renames the vagrant box output with a simpler name in the
vagrant-boxes directory.

Removes SaltStack capability for now, not needed at this time.