Initial chef provisioner with debian pkg
This doesn't have to be final, but currently if the url for the
chef-client .deb pkg is passed in as a variable the build will install
chef-client with no configuration included. The config will still occur
inside of EC2 user data, or elsewhere.

Along with the deb package url, a sha1 checksum variable must be passed
as well to ensure that curl is securely and successfully downloading the
package.

This commit also adds a required_packages function to establish a place
where the base ubuntu packages can be installed identically across each
image type. The actual package listing for each image still needs to be
diffed and more packages will likely need to be installed in the
function.