simplify so only works for vagrant on virtualbox
The real objective is to support custom building of a local dev
environment with the latest Ubuntu. Docker and AWS are out of scope and
just get in the way here.

These changes enable pushing a virtualbox vagrant box to Atlas so that a
user can use `vagrant init username/ubuntu-base` and it will work.